requirements.txt
requests
schedule
fastapi
uvicorn
pyserial
apscheduler